refactor fight from tick to fish
